first (kind of) working version. adding game logic and removing bin folder.

+++ include/config_handler.h
@@ -6,5 +6,5 @@
 
 void load_configs();
 bool config_exists(const char*);
-bool set_grid_positions(uint8_t**, int32_t);
+bool set_grid_positions(uint8_t**, int8_t);
 void free_cfg_resources();
